feat: use long long int but not resolved(because of power...?)

+++ cpp/abc336/c.cpp
@@ -4,8 +4,8 @@
 
 using namespace std;
 
-long int ipow(int base, int exp) {
-    long int result = 1;
+long long int ipow(int base, int exp) {
+    long long int result = 1;
     while (exp > 0) {
         if (exp % 2 == 1) {
             result *= base;
@@ -16,8 +16,8 @@ long int ipow(int base, int exp) {
     return result;
 }
 
-long int create_quinary(long int n) {
-    long int n5 = 0;
+long long int create_quinary(long long int n) {
+    long long int n5 = 0;
     int i = 0;
     while (n > 0) {
         n5 += (n % 5) * ipow(10, i);
@@ -30,11 +30,11 @@ long int create_quinary(long int n) {
 vector<int> Vec = {0, 2, 4, 6, 8};
 
 int main() {
-    long int n;
+    long long int n;
     cin >> n;
-    long int n5 = create_quinary(n-1);
+    long long int n5 = create_quinary(n-1);
     int n5_len = 0;
-    long int tmp = n5;
+    long long int tmp = n5;
     while (tmp > 0) {
         n5_len++;
         tmp /= 10;
